fix(ui): address review feedback on noodles archive
- revert out-of-scope blog Markdown include change in modules/blog.ts
- remove unreferenced public/extra/npmx-sticker.png stray asset
- replace match[1]! non-null assertion with an explicit guard
- extract shared fetchBlueskyAvatars helper reused by blog and noodles
